+++ deployment/v2/refresh-multiple-tests.sh
@@ -74,6 +74,7 @@ aws-lambda-sqs
 aws-lambda-sqs-lambda
 aws-lambda-ssmstringparameter
 aws-lambda-stepfunctions
+aws-lambda-transcribe
 aws-openapigateway-lambda
 aws-route53-alb
 aws-route53-apigateway